requirements.txt
python - binance
TA-Lib
numpy
websocket-client
